sheepdog: change to save glance-image as snapshot
If Cinder and Glance use the same Sheepdog Cluster,
it is cloneable from glance-image to cinder-volume.

Previously Glance-image was saved as Sheepdog volume and
Cinder created the cloned volume after obtaining the snapshot.
Then, Cinder deleted the snapshot.

To reduce the unnecessary calls for obtaining
clone image, this patch changes to save a glance-image as snapshot

Signed-off-by: Teruaki Ishizaki <ishizaki.teruaki@lab.ntt.co.jp>