Sheepdog: improve operations for storing glance-image
This patch changes glance-image to be saved as Sheepdog snapshot.

It reduces operations for creating cinder volume from glance-image
when glance and cinder use the same Sheepdog cluster.

Before:
 - Create a temporary snapshot from volume which is saved
   for glance-image
 - Clone a volume from a temporary snapshot
 - Delete a temporary snapshot

After:
 - Create a volume from the snapshot which is saved
   for glance-image

Please see the clone_image method of the following patch
for cinder driver.
https://review.openstack.org/#/c/202954/

In addition, Sheepdog snapshot is read-only normally and
glance-image cannot be overwritten accidentally.

This patch works without the cinder patch for
Sheepdog driver.

Change-Id: I28ea18badcb4a47daece5efed38d53fc35870902